Move texture_from_palette into a separate lib the binaries can use.
Exclude that package from coverage as it uses calloop on *nix which doesn't work right now for coverage.
