+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
     "./nes/ines_debug",
     "./nes/nes_chr",
     "./nes/nes_pal",
+    "./nes/nes_pal_gui",
     "./nes/nes_pal_render",
 ]
 default-members = [
@@ -43,6 +44,7 @@ default-members = [
     "./nes/ines_debug",
     "./nes/nes_chr",
     "./nes/nes_pal",
+    "./nes/nes_pal_gui",
     "./nes/nes_pal_render",
 ]
 resolver = "2"
@@ -74,6 +76,7 @@ memory = { path = "./memory", version = "~0.1" }
 monitor = { path="./monitor", version = "~0.1" }
 nes_chr = { path="./nes/nes_chr", version = "~0.1" }
 nes_pal = { path="./nes/nes_pal", version = "~0.1" }
+nes_pal_gui = { path="./nes/nes_pal_gui", version = "~0.1" }
 ntest = "0.9.0"
 rand = "0.8.5"
 regex = "1.10.2"
