+++ nes/nes_pal/Cargo.toml
@@ -7,6 +7,4 @@ authors.workspace = true
 
 [dependencies]
 color-eyre = { workspace = true }
-eframe = { workspace = true }
-egui = { workspace = true }
