+++ nes/nes_pal_gui/Cargo.toml
@@ -1,17 +1,13 @@
 [package]
-name = "nes_pal_render"
+name = "nes_pal_gui"
 version = "0.1.0"
 edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
-clap = { workspace = true }
-color-eyre = { workspace = true }
 eframe = { workspace = true }
 egui = { workspace = true }
 nes_pal = { workspace = true }
-nes_pal_gui = { workspace = true }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-env_logger = "0.10"
